+++ cargo-concordium/src/main.rs
@@ -132,7 +132,56 @@ enum Command {
         )]
         module_path:  Option<PathBuf>,
     },
-
+    #[structopt(
+        name = "schema-base64",
+        about = "Convert a schema into its base64 representation and output it to a file or print \
+                 it to the console.
+        A schema has to be provided either as part of a smart contract module or with the schema \
+                 flag. You need to use exactly one of the two flags(`--schema` or `--module`) \
+                 with this command."
+    )]
+    SchemaBase64 {
+        #[structopt(
+            name = "out",
+            long = "out",
+            short = "o",
+            default_value = "-",
+            help = "Path and filename to write the converted base64 representation to or use the \
+                    default value `-` to print the base64 schema to the console. The path has to \
+                    exist while the file will be created. (expected input: \
+                    `./my/path/base64_schema.b64` or `-`)."
+        )]
+        out:          PathBuf,
+        #[structopt(
+            name = "schema",
+            long = "schema",
+            short = "s",
+            conflicts_with = "module",
+            required_unless = "module",
+            help = "Path and filename to a file with a schema (expected input: \
+                    `./my/path/schema.bin`)."
+        )]
+        schema_path:  Option<PathBuf>,
+        #[structopt(
+            name = "wasm-version",
+            long = "wasm-version",
+            short = "v",
+            help = "If the supplied schema or module is the unversioned one this flag should be \
+                    used to supply the version explicitly. Unversioned schemas and modules were \
+                    produced by older versions of `concordium-std` and `cargo-concordium`."
+        )]
+        wasm_version: Option<WasmVersion>,
+        #[structopt(
+            name = "module",
+            long = "module",
+            short = "m",
+            conflicts_with = "schema",
+            required_unless = "schema",
+            help = "Path and filename to a file with a smart contract module (expected input: \
+                    `./my/path/module.wasm.v1`)."
+        )]
+        module_path:  Option<PathBuf>,
+    },
     #[structopt(
         name = "build",
         about = "Build a deployment ready smart-contract module."
@@ -144,14 +193,14 @@ enum Command {
             short = "e",
             help = "Builds the contract schema and embeds it into the wasm module."
         )]
-        schema_embed:    bool,
+        schema_embed:      bool,
         #[structopt(
             name = "schema-out",
             long = "schema-out",
             short = "s",
             help = "Builds the contract schema and writes it to file at specified location."
         )]
-        schema_out:      Option<PathBuf>,
+        schema_out:        Option<PathBuf>,
         #[structopt(
             name = "schema-json-out",
             long = "schema-json-out",
@@ -159,14 +208,24 @@ enum Command {
             help = "Builds the contract schema and writes it in JSON format to the specified \
                     directory."
         )]
-        schema_json_out: Option<PathBuf>,
+        schema_json_out:   Option<PathBuf>,
+        #[structopt(
+            name = "schema-base64-out",
+            long = "schema-base64-out",
+            short = "b",
+            help = "Builds the contract schema and writes it in base64 format to file at \
+                    specified location or prints the base64 schema to the console if the value \
+                    `-` is used. The path has to exist while the file will be created. (expected \
+                    input: `./my/path/base64_schema.b64` or `-`)."
+        )]
+        schema_base64_out: Option<PathBuf>,
         #[structopt(
             name = "out",
             long = "out",
             short = "o",
             help = "Writes the resulting module to file at specified location."
         )]
-        out:             Option<PathBuf>,
+        out:               Option<PathBuf>,
         #[structopt(
             name = "contract-version",
             long = "contract-version",
@@ -174,12 +233,12 @@ enum Command {
             help = "Build a module of the given version.",
             default_value = "V1"
         )]
-        version:         utils::WasmVersion,
+        version:           utils::WasmVersion,
         #[structopt(
             raw = true,
             help = "Extra arguments passed to `cargo build` when building Wasm module."
         )]
-        cargo_args:      Vec<String>,
+        cargo_args:        Vec<String>,
     },
 }
 
@@ -405,67 +464,50 @@ pub fn main() -> anyhow::Result<()> {
                  `./my/path/`)."
             );
 
-            let schema = if let Some(module_path) = module_path {
-                let bytes = fs::read(module_path).context("Could not read module file.")?;
+            let schema = get_schema(module_path, schema_path, wasm_version)
+                .context("Could not get schema.")?;
 
-                let mut cursor = std::io::Cursor::new(&bytes[..]);
-                let (wasm_version, module) = match wasm_version {
-                    Some(v) => (v, &bytes[..]),
-                    None => {
-                        let wasm_version = utils::WasmVersion::read(&mut cursor).context(
-                            "Could not read module version from the supplied module file. Supply \
-                             the version using `--wasm-version`.",
-                        )?;
-                        (wasm_version, &cursor.into_inner()[8..])
-                    }
-                };
+            write_json_schema(&out, &schema).context("Could not write JSON schema files.")?
+        }
+        Command::SchemaBase64 {
+            out,
+            module_path,
+            schema_path,
+            wasm_version,
+        } => {
+            let schema = get_schema(module_path, schema_path, wasm_version)
+                .context("Could not get schema.")?;
 
-                match wasm_version {
-                    utils::WasmVersion::V0 => utils::get_embedded_schema_v0(module).context(
-                        "Failed to get schema embedded in the module.\nPlease provide a smart \
-                         contract module with an embedded schema.",
-                    )?,
-                    utils::WasmVersion::V1 => {
-                        // get the module schema if available.
-                        utils::get_embedded_schema_v1(module).context(
-                            "Failed to get schema embedded in the module.\nPlease provide a smart \
-                             contract module with an embedded schema.",
-                        )?
-                    }
-                }
-            } else if let Some(schema_path) = schema_path {
-                let bytes = fs::read(schema_path).context("Could not read schema file.")?;
-
-                if bytes.starts_with(VERSIONED_SCHEMA_MAGIC_HASH) {
-                    from_bytes::<VersionedModuleSchema>(&bytes)?
-                } else if let Some(wv) = wasm_version {
-                    match wv {
-                        WasmVersion::V0 => from_bytes(&bytes).map(VersionedModuleSchema::V0)?,
-                        WasmVersion::V1 => from_bytes(&bytes).map(VersionedModuleSchema::V1)?,
-                    }
-                } else {
-                    bail!(
-                        "Legacy unversioned schema was supplied, but no version was provided. Use \
-                         `--wasm-version` to specify the version."
+            if out.as_path() == Path::new("-") {
+                write_schema_base64(None, &schema).context("Could not print base64 schema.")?;
+            } else {
+                // A valid path needs to be provided when using the `--out` flag.
+                if out.file_name().is_none() || out.is_dir() {
+                    anyhow::bail!(
+                        "The `--out` flag should point to an existing directory + filename \
+                         (expected input: `./my/path/base64_schema.b64`) or be `-`."
                     );
                 }
-            } else {
-                bail!("Exactly one of `--schema` or `--module` must be provided.");
-            };
 
-            write_json_schema(&out, &schema).context("Unable to output schemas.")?
+                write_schema_base64(Some(out), &schema)
+                    .context("Could not write base64 schema file.")?;
+            }
         }
         Command::Build {
             schema_embed,
             schema_out,
             schema_json_out,
+            schema_base64_out,
             out,
             version,
             cargo_args,
         } => {
             let build_schema = if schema_embed {
                 SchemaBuildOptions::BuildAndEmbed
-            } else if schema_out.is_some() || schema_json_out.is_some() {
+            } else if schema_out.is_some()
+                || schema_json_out.is_some()
+                || schema_base64_out.is_some()
+            {
                 SchemaBuildOptions::JustBuild
             } else {
                 SchemaBuildOptions::DoNotBuild
@@ -526,7 +568,24 @@ pub fn main() -> anyhow::Result<()> {
                          (expected input `./my/path/`)."
                     );
                     write_json_schema(&schema_json_out, module_schema)
-                        .context("Could not write schema file.")?;
+                        .context("Could not write JSON schema files.")?;
+                }
+                if let Some(schema_base64_out) = schema_base64_out {
+                    if schema_base64_out.as_path() == Path::new("-") {
+                        write_schema_base64(None, module_schema)
+                            .context("Could not print base64 schema.")?;
+                    } else {
+                        if schema_base64_out.file_name().is_none() || schema_base64_out.is_dir() {
+                            anyhow::bail!(
+                                "The `--schema-base64-out` flag should point to an existing \
+                                 directory + filename (expected input: \
+                                 `./my/path/base64_schema.b64`) or be `-`."
+                            );
+                        }
+
+                        write_schema_base64(Some(schema_base64_out), module_schema)
+                            .context("Could not write base64 schema file.")?;
+                    }
                 }
                 if schema_embed {
                     eprintln!("   Embedding schema into module.\n");
@@ -1532,6 +1591,65 @@ fn get_parameter(
     }
 }
 
+/// Attempt to get a schema (either from a smart contract module file or a
+/// schema file) from the supplied paths, signalling failure if this is not
+/// possible.
+fn get_schema(
+    module_path: Option<PathBuf>,
+    schema_path: Option<PathBuf>,
+    wasm_version: Option<WasmVersion>,
+) -> anyhow::Result<VersionedModuleSchema> {
+    let schema = if let Some(module_path) = module_path {
+        let bytes = fs::read(module_path).context("Could not read module file.")?;
+
+        let mut cursor = std::io::Cursor::new(&bytes[..]);
+        let (wasm_version, module) = match wasm_version {
+            Some(v) => (v, &bytes[..]),
+            None => {
+                let wasm_version = utils::WasmVersion::read(&mut cursor).context(
+                    "Could not read module version from the supplied module file. Supply the \
+                     version using `--wasm-version`.",
+                )?;
+                (wasm_version, &cursor.into_inner()[8..])
+            }
+        };
+
+        match wasm_version {
+            utils::WasmVersion::V0 => utils::get_embedded_schema_v0(module).context(
+                "Failed to get schema embedded in the module.\nPlease provide a smart contract \
+                 module with an embedded schema.",
+            )?,
+            utils::WasmVersion::V1 => utils::get_embedded_schema_v1(module).context(
+                "Failed to get schema embedded in the module.\nPlease provide a smart contract \
+                 module with an embedded schema.",
+            )?,
+        }
+    } else if let Some(schema_path) = schema_path {
+        let bytes = fs::read(schema_path).context("Could not read schema file.")?;
+
+        if bytes.starts_with(VERSIONED_SCHEMA_MAGIC_HASH) {
+            from_bytes::<VersionedModuleSchema>(&bytes)?
+        } else if let Some(wv) = wasm_version {
+            match wv {
+                WasmVersion::V0 => from_bytes(&bytes).map(VersionedModuleSchema::V0)?,
+                WasmVersion::V1 => from_bytes(&bytes).map(VersionedModuleSchema::V1)?,
+            }
+        } else {
+            bail!(
+                "Legacy unversioned schema was supplied, but no version was provided. Use \
+                 `--wasm-version` to specify the version."
+            );
+        }
+    } else {
+        bail!("Exactly one of `--schema` or `--module` must be provided.");
+    };
+    Ok(schema)
+}
+
+/// Write the JSON representation of the schema into files in the `out`
+/// directory. The files are named after contract_names, except if a
+/// contract_name contains unsuitable characters. Then the counter is used to
+/// name the file.
 fn write_json_schema(out: &Path, schema: &VersionedModuleSchema) -> anyhow::Result<()> {
     match schema {
         VersionedModuleSchema::V0(module_schema) => {